+++ R/map_env_dist.R
@@ -12,6 +12,7 @@
 #'   \item euclidean: Euclidean distance. This metric Z-score standardizes the projection data based on the training data and then standardizes the resulting distance by the maximum pairwise distance within the training data. The result is a similarity value that ranges from 0 to 1.
 #'   \item mahalanobis: Mahalanobis distance. This metric Z-score standardizes the projection data based on the training data and then standardizes the resulting distance by the maximum pairwise distance within the training data. The result is a similarity value that ranges from 0 to 1.
 #'   }
+#' @param n_cores numeric. Number of cores to use for parallel processing when metric is "domain". Default 1 (no parallelization).
 #'
 #' @return
 #' A SpatRaster or tibble object with the nearest environmental distance between presences and projection data.
@@ -53,6 +54,8 @@
 #'
 #' # Measure environmental distance between presences and projection data
 #' clrs = c("#000033", "#1400FF", "#C729D6", "#FF9C63", "#FFFF60")
+#' 
+#' 
 #' # Domain
 #' envdist <-
 #'   map_env_dist(
@@ -74,6 +77,8 @@
 #'   color_p = "red",
 #'   color_gradient = clrs
 #' )
+#' 
+#' 
 #' # Euclidean
 #' envdist <-
 #'   map_env_dist(
@@ -122,7 +127,8 @@
 map_env_dist <- function(
   training_data,
   projection_data,
-  metric = "domain"
+  metric = "domain",
+  n_cores = 1
 ) {
   cell <- y <- x <- cluster <- . <- NULL
 
@@ -157,7 +163,7 @@ map_env_dist <- function(
   }
   if (any("SpatRaster" == class(projection_data))) {
     projection_data <- projection_data[[v0]]
-    extraraster <- projection_data[[1]]
+    extraraster <- projection_data[[which(!is.factor(somevar))[1]]]
     extraraster[!is.na(extraraster)] <- 0
   } else {
     projection_data <- projection_data[v0]
@@ -174,23 +180,22 @@ map_env_dist <- function(
     ncell <- rownames(env_proj2) %>% as.numeric()
   }
 
-  env_proj2 <- env_proj2 %>% dplyr::as_tibble()
-  set <- c(seq(1, nrow(env_proj2), 100), nrow(env_proj2) + 1)
+  env_proj2 <- env_proj2[v0] %>% as.matrix()
+  env_calib2 <- env_calib2[v0] %>% as.matrix()
+
+  set <- c(seq(1, nrow(env_proj2), 1000), nrow(env_proj2) + 1)
 
   extra <- list()
   for (x in seq_len((length(set) - 1))) {
     rowset <- set[x]:(set[x + 1] - 1)
-    if (metric == "domain") {
-      extra[[x]] <- min_gower_rcpp(data1 = env_calib2[v0], data2 = env_proj2[rowset, v0])
-    }
     if (metric == "euclidean") {
       envdist <-
-        euc_dist_stand(env_proj2[rowset, v0], env_calib2[v0])
+        euc_dist_stand(x = env_proj2[rowset, , drop = FALSE], y = env_calib2)
       extra[[x]] <- sapply(data.frame(t(envdist)), min)
     }
     if (metric == "mahalanobis") {
       envdist <-
-        mah_dist_stand(env_proj2[rowset, v0], env_calib2[v0])
+        mah_dist_stand(env_proj2[rowset, , drop = FALSE], env_calib2)
       extra[[x]] <- sapply(data.frame(t(envdist)), min)
     }
   }
@@ -200,9 +205,11 @@ map_env_dist <- function(
     extra <- 1 - extra
     extra[extra < 0] <- 0
   }
-  # extra <- 1 - extra
-  # extra[extra < 0] <- 0
 
+  if (metric == "domain") {
+      extra <- min_gower_rcpp(data1 = env_calib2, data2 = env_proj2, n_threads = n_cores)
+  }
+  
   if (any("SpatRaster" == class(projection_data))) {
     env_proj2 <- data.frame(distance = extra)